+++ client/src/routes.js
@@ -9,20 +9,14 @@ const Team = React.lazy(() => import('./views/components/team/Team'))
 const ShowColumns = React.lazy(() => import('./views/components/columnSummary/showColumns'))
 const Dashboard = React.lazy(() => import('./views/dashboard/Dashboard'))
 const Career = React.lazy(() => import('./views/in/career/Career'))
-const Recommendation = React.lazy(() => import('./views/in/career/recommendation/Recommendation'))
-const Recruitment = React.lazy(() => import('./views/in/career/recruitment/Recruitment'))
-// const AddRecruitment = React.lazy(() =>
-//   import('./views/components/career/addRecruitment/addRecruitment'),
-// )
-// const AddRecommendation = React.lazy(() =>
-//   import('./views/components/career/addRecommendation/addRecommendation'),
-// )
-const EditRecruitment = React.lazy(() => import('./views/in/career/recruitment/EditRecruitment'))
+const RecommendationForm = React.lazy(() => import('./views/in/recommendation/RecommendationForm'))
+const Recommendation = React.lazy(() => import('./views/in/recommendation/Recommendation'))
+const RecruitmentForm = React.lazy(() => import('./views/in/recruitment/RecruitmentForm'))
+const Recruitment = React.lazy(() => import('./views/in/recruitment/Recruitment'))
+const EditRecruitment = React.lazy(() => import('./views/in/recruitment/edit/EditRecruitment'))
 const EditRecommendation = React.lazy(() =>
-  import('./views/in/career/recommendation/EditRecommendation'),
+  import('./views/in/recommendation/edit/EditRecommendation'),
 )
-
-const CareerForm = React.lazy(() => import('./views/in/career/CareerForm'))
 
 const ProfileSearch = React.lazy(() => import('./views/components/searchProfile/SearchProfile'))
 
@@ -34,12 +28,12 @@ const routes = [
   { path: '/career', exact: true, name: 'Career', component: Career },
   { path: '/recommendation', exact: true, name: 'Recommendation', component: Recommendation },
   { path: '/recruitment', exact: true, name: 'Recruitment', component: Recruitment },
-  { path: '/addRecruitment', exact: true, name: 'AddRecruitment', component: CareerForm },
+  { path: '/addRecruitment', exact: true, name: 'AddRecruitment', component: RecruitmentForm },
   {
     path: '/addRecommendation',
     exact: true,
     name: 'AddRecommendation',
-    component: CareerForm,
+    component: RecommendationForm,
   },
   {
     path: '/editRecruitment/:id',
