+++ client/src/views/in/career/index.js
@@ -3,12 +3,5 @@ import { selectCareer, setKeywords, clearKeywords } from '../../../slices/career
 import eesa from '../../../assets/images/eesa-icon.png'
 import Recruitment_image from '../../../assets/images/Recruitment.png'
 import Recommendation_image from '../../../assets/images/Recommendation.png'
-export {
-  Career,
-  selectCareer,
-  setKeywords,
-  clearKeywords,
-  eesa,
-  Recommendation_image,
-  Recruitment_image,
-}
+export { selectCareer, setKeywords, clearKeywords, eesa, Recommendation_image, Recruitment_image }
+export default Career
