+++ client/src/views/in/recommendation/edit/EditRecommendation.js
@@ -1,7 +1,7 @@
 /* eslint-disable prettier/prettier */
 import React, { useState, useEffect } from 'react'
 import { useParams } from 'react-router-dom'
-import CareerForm from '../CareerForm'
+import CareerForm from '../../career/CareerForm'
 import axios from 'axios'
 const EditRecommendation = () => {
   const id = useParams().id
